Update networkMaster.py
It appeared that I was still missing some resent packets so I began to examine how my resent packets were being counted. Rather than count them outwith the function that calls them, I decided to add a global resentPackets list and add each resent packet to this list after sending to it's relevant node. 

At the end of the iteration I then count the length of the resentPackets list and add that to 6012 (the total number of packets created). This gives me the magic number that the total packets in the forwardList, the packetList and the bufferList must equal. 

After counting how any resent packets have actually been sent in this way all my numbers now add up.