Update networkMaster.py
Continuing to add functionality to my network my plan was to find out the fastest routes for any particular journey on the network. The above was an experimental approach to the problem using dictionaries. 

The essential data was the time it took to arrive at its destination and the next node in the network the packet had visited. I set up a dictionary using the times as the key and the node number as the value and the used the collections library to produce an ordered tuple list of best times. 

This approach was very fast to code but came with two major flaws. Firstly, the output is very hard to make sense off as the conversion to an 'orderedDictionary' is taking place within the FOR loop. Secondly. but most importantly it takes almost a full 6 seconds to provide an answer, and for a network this small this is completely unacceptable. 

I have solved the matter of producing some best results for every journey in another way, using a 3d array, and have merged from another branch.