Update networkMaster.py
Continuing to add more functionality to the network I wanted to find out what the best route any particular packet can take on any particular journey around the network. This information will then be fed back into my network so that the system can automatically adjust it's routing table. 

I attempted an experimental way of solving this problem using Dictionaries to store and sort my information, however this involved putting a lot of the computational power onto the system and therefor took an unacceptable time to provide an answer. 

This more traditional method of using a 2-dimensional array took only slightly longer to code but produced vastly optimised results. 

An exception statement to catch two routes that go through the same 'best' node also needs to be added so that we get a more accurate choice when deciding which route to send a particular packet.