Update networkMaster.py
Update: some minor formatting to better see the quickest journey's through the network. Journey 1 represents from Node 1 to Node 1 (hence why it is populated by (0,0,0) and (999,999,999), representing a journey which does not exist), journey 2 is going from Node 1 to Node 2, journey 3 is going from Node 1 to Node 3 and so on. 

Future update: Using the data extracted from the network I will build a function to automatically adjust the routing table to choose the optimum 'next' node to send any packet on any particular journey.