Update networkMaster.py
Update fix: The table at the end was only populating index [0] with one 'best' journey. The problem was the code that made the comparison between a time a packet has arrived and my top three best time list. 

To solve this issue I now move everything along by one if the time a packet has arrived is shorter than the first time in the best time list. 

The system now outputs the three quickest packets to have made a particular journey, the time it took to make that journey and the first Node the packet visited when making that journey. 

Future update: With the lists now properly populated I can write a function to automatically update a node's routing table based on the best journeys around the network.