Update networkMaster.py
Update: my topThree function was not properly comparing and shifting items in my bestTime list. This has now been fixed. 

Future Update: Using the information from my bestTimes list I will create a function that automatically updates a nodes routing table. 

Also considered that it may be better to hold the routing table in a dictionary (as I want to associate a list of particular nodes with a certain address). This will allow for faster lookup time rather than iterating over a node's routing table in a regular Python array.