Update networkMaster.py
Update: I changed the routing table from a set of three lists to a dictionary. This has meant my node object need less data to initialise, and I also no longer have to iterate over a set of lists to provide a routing table. The final benefit is that I no longer have to check for 0's within a routing table (i.e when a route did not exist). 

Future Update: My top journey's need to compare with each other, and if the route is the same, do not update. I am hoping this will provide some different 'next best nodes' to visit within my set of top three tuples.