Update networkMaster.py
Bug Fix: The network was sticking every now and then while trying to delete streaming packets. The check_count was not counting properly and packets were not being deleted so the program was stuck inside the WHILE loop.  I deleted that whole block of code and re-wrote the section, which now deletes packets properly from the checking list and counts down the check_count properly, exiting from the WHILE loop and continuing to the next node. 

With this bug solved I attempted to iterate over the network, but another bug revealed itself in that when packets were being re-sent, I was hitting a listIndexError when deleting from the sent_packets list. Again, I just deleted that small block of code and re-wrote in a more pythonic way. This solved the problem. 

The network now iterates five times and provides fastest journeys for all possible journeys on the network. 

Future Update: The network really slows down and I suspect this is because lists are not being emptied properly after each iteration. I aim to speed the network up, then have the system assess the best journeys and adjust each nodes routing table accordingly.