Update networkMaster.py
Bug Fix: Even after removing all my duplicates I was still having a few 'top journeys' delivering 0.0 times, with those packets having visited multiple nodes, multiple times before arriving. 

Using my packet.resent flag I was able to work out that all the packets delivering 0.0 times and multiple node visits were all resent. I added a small statement to my find_fastest_three() function that checks if a packet has been resent, and if so it breaks from that iteration of the loop. 

This makes sense, as a packet that was resent will not arrive sooner than a packet that was not resent. 

After this adjustment I now have correct times for each journey and the best node on which to begin the journey.