Update networkMaster.py
Update: Having fixed the bug where I was receiving odd best_times in my best_time_list, I have now begun putting together the functionality so the network can update its routing paths based on the next_best_node lists produced by the best_time function. 

The program will effectively run for a number of iterations, after which the packets will be tested, the best_node_lists acquired and the whole networks routing system updated. This is achieved with a call to update_routing within the node objects themselves. 

Future Update: I need to remove duplicate best_nodes from the next_best_node lists. I plan to achieve this using a seperate function (also within the node object) that can reference the nodes original linkage list, and suggest next_best_nodes based on that (provided there are duplicates or 0's (a flag that indicates a journey was no assessed / completed).

After this the plan is to create a genetic topology generator that builds a network, runs the tests and provides output on a certain topologies speeds. This data will be used to genetically evolve an optimum network (given parameters such as fibre link limitations, number of nodes and cost).