Update networkMaster.py
Update: The network now runs for ten iterations, tests the speed of all the packets sent and updates all the nodes routing tables with the next "best node" to travel to on a particular journey. 

Future Update: Implement a "Test" switch to go from a "Testing" state to a "Running" state after measuring the speed of packets moving through the network.