Update networkMaster.py
Update: The network will now run for a number of test iterations, after which it will calculate the top three times and the relevant nodes that were visited to get to those times. After this my active_network switch is turned to True, the network examines these best times and creates a "Top Next Node" routing table within each node object. This is then used when the network runs for a number of "running iterations", after which the top times are printed out again. 

Future Update: Next up, genetic topology generator.